updates repos

+++ src/main/java/lk/ijse/gdse/traditionalexperiencebackend/repo/WorkshopRepo.java
@@ -3,5 +3,5 @@ package lk.ijse.gdse.traditionalexperiencebackend.repo;
 import lk.ijse.gdse.traditionalexperiencebackend.entity.Workshop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkshopRepo extends JpaRepository<Workshop, Integer> {
+public interface WorkshopRepo extends JpaRepository<Workshop, Long> {
 }
